template/fastapi/requirements.txt
fastapi>=0.104.0
hypercorn==0.14.4
uvicorn>=0.24.0
mlflow>=2.8.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.4.2
python-multipart>=0.0.6
